refactor(desktop): split workspace app context IPC
